requirements.txt
django==2.0.3
mysqlclient==1.3.12
django-storages==1.6.5
boto3==1.6.11
pillow==5.0.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
django-oauth-toolkit==1.0.0
django-cors-headers==2.2.0